Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
torch==2.2.0
torchvision==0.17.0
transformers==4.35.2
huggingface-hub==0.19.4
numpy==1.26.2
python-dotenv==1.0.0
pydantic==2.5.0
